update type model
