+++ core/src/main/java/org/nervos/ckb/methods/response/CkbCell.java
@@ -1,7 +1,7 @@
 package org.nervos.ckb.methods.response;
 
 import org.nervos.ckb.methods.Response;
-import org.nervos.ckb.methods.response.item.Cell;
+import org.nervos.ckb.methods.type.Cell;
 
 /**
  * Created by duanyytop on 2018-12-25.
