+++ core/src/main/java/org/nervos/ckb/methods/response/CkbTransaction.java
@@ -1,7 +1,7 @@
 package org.nervos.ckb.methods.response;
 
 import org.nervos.ckb.methods.Response;
-import org.nervos.ckb.methods.response.item.Transaction;
+import org.nervos.ckb.methods.type.Transaction;
 
 /**
  * Created by duanyytop on 2019-01-07.
