+++ core/src/main/java/org/nervos/ckb/methods/type/Output.java
@@ -1,0 +1,21 @@
+package org.nervos.ckb.methods.type;
+
+/**
+ * Created by duanyytop on 2019-01-08.
+ * Copyright © 2018 Nervos Foundation. All rights reserved.
+ */
+public class Output {
+
+    /**
+     * capacity : 50000
+     * type : null
+     * data : 0x
+     * lock : 0x321c1ca2887fb8eddaaa7e917399f71e63e03a1c83ff75ed12099a01115ea2ff
+     */
+
+    public long capacity;
+    public Object type;
+    public String lock;
+    public String data;
+
+}
